requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
openai>=1.12.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
